pipelines/rag_ranker/pyproject.toml
[tool.poetry]
name = "rutube-qna-pipeline-rag-ranker"
version = "0.1.0"
description = "Rutube Q&A Pipeline RAG Ranker"
license = "Apache-2.0"
authors = ["AI RnD Lab Authors"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
haystack-ai = "^2.5.1"
datasets = ">=2.6.1"
sentence-transformers = ">=3.0.0"
pymorphy2 = "^0.9.1"
catboost = "^1.2.7"
nltk = "^3.9.1"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
